Change gettng the fnref logic
