+++ util/ivmFactory.js
@@ -6,6 +6,14 @@ const {
   getLibraryCode
 } = require("./customTransforrmationsStore");
 
+const compiledModules = {};
+
+async function loadModule(isolate, context, moduleName, moduleCode) {
+  const module = await isolate.compileModule(moduleCode);
+  await module.instantiate(context, () => {});
+  compiledModules[moduleName] = { module };
+}
+
 async function createIvm(versionId, libraryVersionIds) {
   const transformation = await getTransformationCode(versionId);
   // libraryVersionIds = [
@@ -17,7 +25,8 @@ async function createIvm(versionId, libraryVersionIds) {
       getLibraryCode(libraryVersionId)
     )
   );
-  if (transformation && libraries) { //TODO: Check if this should this be &&
+  if (transformation && libraries) {
+    //TODO: Check if this should this be &&
     const librariesMap = {};
     libraries.forEach(library => {
       librariesMap[_.camelCase(library.name)] = library.code;
@@ -36,7 +45,7 @@ function metadata(eventsMetadata, event) {
     return eventsMetadata[event.messageId] || {};
 }
 
-function transform(fullEvents) {
+export function transform(fullEvents) {
   const events = fullEvents.map(event => event.message);
   const eventsMetadata = {};
   fullEvents.forEach(ev => {
@@ -216,20 +225,21 @@ function transform(fullEvents) {
 
   await Promise.all(
     supportedFuncNames.map(async sName => {
-      const funcRef = await jail.get(sName);
+      const funcRef = await customScriptModule.namespace.get(sName);
       if (funcRef) {
         supportedFuncs.push(funcRef);
       }
     })
   );
 
-  if (supportedFuncs.length !== 1) {
-    throw new Error(
-      `Expected one of ${supportedFuncNames}. Found ${supportedFuncs.map(
-        sFunc => sFunc.name
-      )}`
-    );
-  }
+  // TODO: Figutre out how to check for the correct function name
+  // if (supportedFuncs.length !== 1) {
+  //   throw new Error(
+  //     `Expected one of ${supportedFuncNames}. Found ${supportedFuncs.map(
+  //       sFunc => sFunc.name
+  //     )}`
+  //   );
+  // }
 
   const fnRef = supportedFuncs[0];
   // TODO : check if we can resolve this
@@ -239,11 +249,11 @@ function transform(fullEvents) {
 
 async function getFactory(versionId, libraryVersionIds) {
   const factory = {
-    create: () => {
+    create: async () => {
       return createIvm(versionId, libraryVersionIds);
     },
-    destroy: client => {
-      client.isolate.dispose();
+    destroy: async client => {
+      await client.isolate.dispose();
     }
   };
